requirements.txt
asgiref==3.5.0
Deprecated==1.2.13
Django==4.0.2
django-redis==5.2.0
djangorestframework==3.13.1
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.4
six==1.16.0
sqlparse==0.4.2
wrapt==1.13.3
